requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.64.0
tiktoken>=0.4.0
google-generativeai>=0.3.0
